requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
audioop-lts==0.2.1
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.12
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.11
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.2
Deprecated==1.2.18
distro==1.9.0
docopt==0.6.2
durationpy==0.10
faiss-cpu==1.11.0
fastapi==0.115.9
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.6.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.32.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
Js2Py==0.74
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.25
langchain-core==0.3.63
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.8
langsmith==0.3.43
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mmh3==5.1.0
mpmath==1.3.0
multidict==6.4.4
narwhals==1.41.0
numpy==2.2.6
oauthlib==3.2.2
onnxruntime==1.22.0
opencv-python==4.12.0.88
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
orjson==3.10.18
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pipwin==0.5.2
posthog==4.2.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyjsparser==2.7.1
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
PyPika==0.48.9
pypiwin32==223
PyPrind==2.11.3
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyttsx3==2.98
pytz==2025.2
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
selenium==4.33.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.45.3
streamlit==1.45.1
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.1
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
twilio==9.6.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
undetected-chromedriver==3.5.5
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
watchdog==6.0.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0
